% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessMh.R
\name{assessMh}
\alias{assessMh}
\title{This function asseses some frequentist properties when naively fitting Mb to capture histories that are missing data.}
\usage{
assessMh(N, p, occ, K, traptype, lambda_h, lambda_c = NULL, delta = NULL,
  kappa_h = NULL, kappa_t = NULL, cluster = FALSE, kappa_c = NULL,
  alpha = NULL, hettype = "logitnormal", sims = 100, cores = 1)
}
\arguments{
\item{N}{an integer that is the true population size}

\item{p}{a vector of size 2 containing the capture and recapture probabilities}

\item{occ}{an integer that is the number of capture occasions to simulate}

\item{K}{an integer that is the number of traps to simulate}

\item{traptype}{a character string specifying the trap type.  "single"
allows individuals to be caught in at most 1 trap per occasion and "multi" allows
individuals to be caught in multiple traps per occasion.}

\item{lambda_h}{a positive value specifying the parameter for the zero-truncated Poisson
hair deposition process (S_ijk|W_ijk=1)}

\item{lambda_c}{a positive value specifying the parameter for the zero-truncated Poisson
cluster deposition process if cluster=TRUE}

\item{delta}{a numeric value between 0 and 1 specifying the probability a hair sample will be retained
in the subsample.  This is used if subsampling is done by pooling samples on each occasion and taking
a simple random sample.}

\item{kappa_h}{an integer specifying the number of hair samples to retain per occasion-trap if cluster=FALSE
or per occasion-trap-cluster if cluster=TRUE.  Used in sub-hair and sub-cluster methods}

\item{kappa_t}{an integer specifying the number of traps per occasion at which to retain kappa_h hair samples.
If no value is specified no trap-level subsampling occurs.  Only works when cluster=FALSE.  Used in the
sub-trap method}

\item{cluster}{a logical indicating whether or not to simulate from the cluster model (multiple clusters per
individual-occasion-trap)}

\item{kappa_c}{an integer specifying the number of clusters per trap-occasion at which to retain kappa_h hair samples.
Only works when cluster=TRUE.  Sub-cluster method.}

\item{alpha}{a numeric value between 0 and 1 specifying the probability that a hair sample produces
an individual identification}

\item{hettype}{...}

\item{sims}{the number of simulated data sets to generate and fit Mb to}

\item{cores}{the number of cores to do the simulation on.  A typical for loop is used if cores=1 and
the foreach package is used if cores>1}
}
\value{
a list with the simulation results (expand later)
}
\description{
This function asseses some frequentist properties when naively fitting Mb to capture histories that are missing data.
}

