% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSS.R
\name{calcSS}
\alias{calcSS}
\title{This function calculates the failure criterion for Mb (see Otis 1978) from the observed capture history Wobs2D
and the sufficient statistics for Mb from both the true and observed capture histories.  (add description of SS later)}
\usage{
calcSS(W2D, Wobs2D, Wobsfull2D)
}
\arguments{
\item{Wobs2D}{the observed capture history (after subsampling and DNA amplification) of dimension N x occ}

\item{W}{the true capture history of dimension N x occ}
}
\value{
a list containing the sufficient statistics and failure criterion
}
\description{
This function calculates the failure criterion for Mb (see Otis 1978) from the observed capture history Wobs2D
and the sufficient statistics for Mb from both the true and observed capture histories.  (add description of SS later)
}

