% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSUR.R
\name{calcSUR}
\alias{calcSUR}
\title{This function calculates the failure criterion for Mb (see Otis 1978) from the observed capture history Wobs2D
and the sufficient statistics for Mb from both the true and observed capture histories.  (add description of SS later)}
\usage{
calcSUR(S2D, U2D, R2D)
}
\arguments{
\item{S}{the true hair capture history of dimension N x occ}

\item{U}{the hair capture history after subsampling of dimension N x occ}

\item{R}{the hair capture history after subsampling and DNA Identificationof dimension N x occ}
}
\value{
a list containing the stuff
}
\description{
This function calculates the failure criterion for Mb (see Otis 1978) from the observed capture history Wobs2D
and the sufficient statistics for Mb from both the true and observed capture histories.  (add description of SS later)
}

