% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capture}
\alias{capture}
\title{This function simulates the capture process}
\usage{
capture(p, N, occ, traptype, btype, K = NULL)
}
\arguments{
\item{p}{a vector of size 2 containing the capture and recapture probabilities}

\item{N}{an integer that is the true population size}

\item{occ}{an integer that is the number of capture occasions to simulate}

\item{traptype}{a character string specifying the trap type.  "single" 
allows individuals to be caught in at most 1 trap per occasion and "multi" allows
individuals to be caught in multiple traps per occasion.}

\item{btype}{a character string specifying the type of behavioral response to capture
Current options are "global" for a global trap response and "trap" for a trap-specific
response.  "trap" currently only works with traptype=multi.}

\item{K}{an integer that is the number of traps to simulate if traptype="multi"}
}
\value{
W a capture history of dimension N x occ if traptype="single" or 
N x occ x K if traptype="multi"
}
\description{
This function simulates the capture process
}

