% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.spatial.R
\name{capture.spatial}
\alias{capture.spatial}
\title{This function simulates a spatial capture-recapture process with the gaussian hazard detection function}
\usage{
capture.spatial(N, occ, lambda, sigma, locs, buff, btype = NULL)
}
\arguments{
\item{N}{an integer that is the true population size}

\item{occ}{an integer that is the number of capture occasions to simulate}

\item{lambda}{lambda rate parameter at 0 for hazard gaussian detection function}

\item{sigma}{the spatial scale parameter}

\item{locs}{a J x 2 matrix of trap locations}

\item{buff}{an integer specifying the buffer for simulating the population}

\item{btype}{a character string specifying the type of behavioral response to capture
Current options are "global" for a global trap response and "trap" for a trap-specific
response.  "trap" currently only works with traptype=multi.}
}
\value{
W a capture history of dimension N x occ x K ...
}
\description{
This function simulates a spatial capture-recapture process with the gaussian hazard detection function
}

