% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captureMh2.R
\name{captureMh2}
\alias{captureMh2}
\title{This function simulates the capture process with a 2-component finite mixture model for individual heterogeneity.}
\usage{
captureMh2(p, N, occ, traptype, K = NULL)
}
\arguments{
\item{p}{a vector of size 3 containing the two capture probabilities and the mixture proportion  p=c(p1,p2,mix)}

\item{N}{an integer that is the true population size}

\item{occ}{an integer that is the number of capture occasions to simulate}

\item{traptype}{a character string specifying the trap type.  "single" 
allows individuals to be caught in at most 1 trap per occasion and "multi" allows
individuals to be caught in multiple traps per occasion.}

\item{K}{an integer that is the number of traps to simulate if traptype="multi"}
}
\value{
W a capture history of dimension N x occ if traptype="single" or 
N x occ x K if traptype="multi"
}
\description{
This function simulates the capture process with a 2-component finite mixture model for individual heterogeneity.
}

