% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deposit.R
\name{deposit}
\alias{deposit}
\title{This function simulates the hair deposition process}
\usage{
deposit(W, lambda_h, traptype = "multi", K, cluster = FALSE,
  lambda_c = NULL)
}
\arguments{
\item{W}{a capture history of dimension N x occ if traptype="single" or 
N x occ x K if traptype="multi"}

\item{lambda_h}{a positive value specifying the parameter for the zero-truncated Poisson
hair deposition process (S_ijk|W_ijk=1)}

\item{traptype}{a character string specifying the capture process.  "single" 
allows individuals to be caught in at most 1 trap per occasion and "multi" allows
individuals to be caught in multiple traps per occasion.  If traptype="single", a trap id is randomly 
assigned to each capture}

\item{K}{an integer that is the number of traps to simulate}

\item{cluster}{a logical indicating whether or not to simulate from the cluster model (multiple clusters per
individual-occasion-trap)}

\item{lambda_c}{a positive value specifying the parameter for the zero-truncated Poisson
cluster deposition process for the cluster model}
}
\value{
S a "hair capture history" containing the number of hair samples deposited by individual i on occasion
j in trap k and if for cluster=TRUE, at cluster l.  S is of dimension N x occ x K if cluster=FALSE and dimension
N x occ x K x Lmax if cluster=TRUE, where Lmax is the maximum observed number of clusters deposited across
trap-occasions
}
\description{
This function simulates the hair deposition process
}

