% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVHyper.R
\name{dMVhyper}
\alias{dMVhyper}
\title{PDF for the multivariate hypergeometric distribution}
\usage{
dMVhyper(k, K)
}
\arguments{
\item{k}{adf}

\item{K}{adf}
}
\description{
PDF for the multivariate hypergeometric distribution
}

