% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAID.R
\name{DNAID}
\alias{DNAID}
\title{This function simulates the DNA identification process}
\usage{
DNAID(alpha, U)
}
\arguments{
\item{alpha}{a numeric value between 0 and 1 specifying the probability that a hair sample produces
an individual identification}

\item{U}{a "subsampled hair capture history" of dimension N x occ x K if cluster=FALSE or 
N x occ x K x Lmax if cluster=TRUE}
}
\value{
R an "observed hair capture history" containing the number of hair samples deposited by individual i on occasion
j in trap k and if for cluster=TRUE, at cluster l, that are retained in the subsample and produce an individual
identification.  R is of dimension N x occ x K if cluster=FALSE and dimension N x occ x K x Lmax if cluster=TRUE,
where Lmax is the maximum observed number of clusters deposited across trap-occasions
}
\description{
This function simulates the DNA identification process
}

