\name{Mbmiscap-package}
\alias{Mbmiscap-package}
\alias{Mbmiscap}
\docType{package}
\title{
Mb with missing data

}
\description{
This package simulates from closed population (classical and spatial) mark recapture models where individuals are captured using hair snares and data may be missing due to subsampling and/or failed DNA amplification.  It also fits the models simulated from without accounting for the subsampling or DNA amplification processes to assess bias.

}
\details{
\tabular{ll}{
Package: \tab Mbmiscap\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-31\cr
License: \tab GPL (>=2) \cr
}
details
}
\author{
Ben Augustine

Maintainer: Ben Augustine <baugusti@vt.edu>
}
