% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVHyper.R
\name{rMVhyper}
\alias{rMVhyper}
\title{Random generator for the multivariate hypergeometric distribution}
\usage{
rMVhyper(n, K)
}
\arguments{
\item{n}{ads}

\item{K}{adf}
}
\description{
Random generator for the multivariate hypergeometric distribution
}

