% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSpatial.R
\name{simSpatial}
\alias{simSpatial}
\title{This function simulates a spatial hair snare mark recapture experiment with hair sample subsampling and
failed DNA identification.}
\usage{
simSpatial(N, occ, lambda, sigma, locs, buff, btype, lambda_h,
  lambda_c = NULL, delta = NULL, kappa_h = NULL, kappa_t = NULL,
  cluster = FALSE, kappa_c = NULL, alpha = NULL)
}
\arguments{
\item{N}{an integer that is the true population size}

\item{occ}{an integer that is the number of capture occasions to simulate}

\item{lambda}{lambda rate parameter at 0 for hazard gaussian detection function}

\item{sigma}{the spatial scale parameter}

\item{locs}{a J x 2 matrix of trap locations}

\item{buff}{an integer specifying the buffer for simulating the population}

\item{btype}{a character string specifying the type of behavioral response to capture
Current options are "global" for a global trap response and "trap" for a trap-specific
response.  "trap" currently only works with traptype=multi.}

\item{lambda_h}{a positive value specifying the parameter for the zero-truncated Poisson
hair deposition process (S_ijk|W_ijk=1)}

\item{lambda_c}{a positive value specifying the parameter for the zero-truncated Poisson
cluster deposition process if cluster=TRUE}

\item{delta}{a numeric value between 0 and 1 specifying the probability a hair sample will be retained
in the subsample.  This is used if subsampling is done by pooling samples on each occasion and taking
a simple random sample.}

\item{kappa_h}{an integer specifying the number of hair samples to retain per occasion-trap if cluster=FALSE
or per occasion-trap-cluster if cluster=TRUE.  Used in sub-hair and sub-cluster methods}

\item{kappa_t}{an integer specifying the number of traps per occasion at which to retain kappa_h hair samples.
If no value is specified no trap-level subsampling occurs.  Only works when cluster=FALSE.  Used in the
sub-trap method}

\item{cluster}{a logical indicating whether or not to simulate from the cluster model (multiple clusters per
individual-occasion-trap)}

\item{kappa_c}{an integer specifying the number of clusters per trap-occasion at which to retain kappa_h hair samples.
Only works when cluster=TRUE.  Sub-cluster method.}

\item{alpha}{a numeric value between 0 and 1 specifying the probability that a hair sample produces
an individual identification}
}
\value{
a list with the simulated quantities and statistics (expand description later)
}
\description{
This function simulates a spatial hair snare mark recapture experiment with hair sample subsampling and
failed DNA identification.
}

