% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{This function simulates the subsampling process}
\usage{
subsample(S, Sdot, delta = NULL, kappa_h = NULL, kappa_t = NULL,
  cluster = FALSE, kappa_c = NULL)
}
\arguments{
\item{S}{a "hair capture history" of dimension N x occ x K if cluster=FALSE or 
N x occ x K x Lmax if cluster=TRUE}

\item{Sdot}{a matrix of constraints on S obtained by summing S across the second and third dimensions.
apply(S,2:3,sum)  Sdot will be 2 dimensional if cluster=FALSE and 3 dimensional if cluster=TRUE}

\item{delta}{a numeric value between 0 and 1 specifying the probability a hair sample will be retained
in the subsample.  This is used if subsampling is done by pooling samples on each occasion and taking
a simple random sample.}

\item{kappa_h}{an integer specifying the number of hair samples to retain per occasion-trap if cluster=FALSE
or per occasion-trap-cluster if cluster=TRUE.  Used in sub-hair and sub-cluster methods}

\item{kappa_t}{an integer specifying the number of traps per occasion at which to retain kappa_h hair samples.
If no value is specified no trap-level subsampling occurs.  Only works when cluster=FALSE.  Used in the
sub-trap method}

\item{cluster}{a logical indicating whether or not to simulate from the cluster model (multiple clusters per
individual-occasion-trap)}

\item{kappa_c}{an integer specifying the number of clusters per trap-occasion at which to retain kappa_h hair samples.
Only works when cluster=TRUE.  Sub-cluster method.}
}
\value{
U a "subsampled hair capture history" containing the number of hair samples deposited by individual i on occasion
j in trap k and if for cluster=TRUE, at cluster l, that are retained in the subsample.  U is of dimension N x occ x K if cluster=FALSE and dimension
N x occ x K x Lmax if cluster=TRUE, where Lmax is the maximum observed number of clusters deposited across
trap-occasions
}
\description{
This function simulates the subsampling process
}

